adds datapoints values as float value

+++ src/datapoints.in.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Datapoints {
     name: String,
-    datapoints: Vec<Vec<i64>>,
+    datapoints: Vec<Vec<f64>>,
     tags: HashMap<String,String>,
     ttl: u32
 }
@@ -18,8 +18,8 @@ impl Datapoints {
         }
     }
 
-    pub fn add(&mut self , ms: i64, value: i64) {
-        self.datapoints.push(vec![ms, value]);
+    pub fn add(&mut self , ms: i64, value: f64) {
+        self.datapoints.push(vec![ms as f64, value]);
     }
 
     pub fn add_tag(&mut self, name: &str, value: &str) {
